Update battery_management_system.py
This is the code for the battery management system. It uses the ADS1115 Analog to Digital Converter and reads the voltage values of Analog pin A0. 

Unfortunately I do not have access to an ADS1115 to test the code so it will have to be verified by testing it as per the instructions before use.